Chore (Manufacturers): Change the functionalitie of the manufacturers from the back and front
